Extract byId.js and define getTodo function
